New GLSL->HLSL cross compiler branch
We are working on a GLSL to HLSL cross-compiler which is based on glslang. Out glslang fork is under https://github.com/nbenty/glslang. We use Falcor as a test bed. The ultimate goal is to have a shading language abstraction to simplify our internal shaders development.
Changes in this commit:
The glslang front-end is missing some features, and is also much more strict than the GLSL compiler in the driver, so some of our common shader code uses idioms that don't compile under glslang.
    - Parameters with default values not supported. I worked around this by adding overloads with fewer parameters that call the original function with the default values.
    - Pointers not supported (no GL_NV_shader_buffer_load). I just #ifdef'd the code out.
    - I hacked glslang to allow `uintptr_t` as an alias for `uint64`, but the version of HLSL I'm generating doesn't support 64-bit integers, so for now I changes the `uintptr_t ptr` member of `BufPtr` over to a `uint[2]` array.
    - Incorrect semantics for `uvec2` comparison. We have one spot where we do `any(uvec2(samp) != 0)`, which is actually incorrect in two ways:
      	- GLSL won't implicitly convert the integer literal `0` to type `uvec2`, so we should be explicitly comparing to `uvec2(0)`
	- GLSL's vector `==` and `!=` are *not* implicitly component-wise (like in HLSL), so the result of `uvec2(samp) != uvec2(0)` is an ordinary `bool`, and not a `bvec2`. The use of `any()` is uncessary, and in strict GLSL it is invalid.
       The correct code would be `uvec2(samp) != uvec2(0)`, but I don't actually use that since we are converting to HLSL which doesn't have bindless anyway, so I just `return true` and have done. :)

Other changes:
- New helper class to create Process and handle its IO.
- Define FALCOR_GLSL even when we are on DX. This means that HLSL is not supported for now.
- Invoke the cross compiler when in DX11 mode.

NOTE: THE CROSS-COMPILER BINARY WAS NOT COMMITED. Git binary limitations. You'll need to build the compiler yourself and put it under Framework/Externals.
Note 2: That doesn't work yet. TextRenderer.vs fails because we are using sampler2D inside a uniform buffer. WIP.