Put cross-compiler build directly into Falcor.
This currently assumes that the appropriate branch (currently: `falcor-dev`) of `glslang` is located under `Externals/`. The main Falcor project then directly builds the source files that we need (or rather, a conservative super-set of them).

The source of the cross-compiler itself is moved into the Falcor build, and changed from a standalone .exe into a relatively trivial library. Right now the code isn't taking full advantage of being inside the Falcor framework/namespace, so it can probably be streamlined further with a bit of time and effort.