Better abstraction for command list/queue/allocator.
The RenderContext class (which I'll probably rename) abstracts the graphics command submission. It creates a single command list and n command allocators.
The Device class holds a single command-queue. Before presenting, it will close and submit the command queue.

This change made the CommandList class redundant, so I removed it.