Work towards HLSL->GLSL cross-compilation.
This change just includes the edits to the core cross-compilation infrastructure, so there is no code to test it in this change.

A lot of stuff still needs to be implemented and/or cleaned up, but the basic translation is here.
That is: I have an HLSL version of the ModelViewer.fs shader compiling, generating HLSL, and loading.
That includes parsing a bunch of code in the common files (I have a hack-job "port" of those to support HLSL in another branch).
It links against the GLSL vertex shader just fine (all uniform-buffer matching passes).
The remaining issue is that it doesn't actually yield the correct result (I just get a garbage constant back), but that is actually a smaller issue in the grand scheme of things. :)