This branch is dedicated for Armor, the little red tank.
I had to remove vertex colors from the defaultVS because Armor doesn't use them. Apparently, D3D12 excepts the VS input signature to fully match the input-layout. This isn't the same behavior as D3D11, so that sucks. I don't have a solution, so for the sake of Armor I removed vertex colors.
Nothing crashes, but I get an error that the root signature doesn't match the shaders. Weird, given that I create the root signature based on the program reflection.