Minor design change - PipelineStateCache is now PipelineState. It holds everything required for rendering, not just the data that is used to create PipelineStateObject.
This mainly means that RenderContext doesn't really work with state directly - it accepts ProgramVars and PipelineState and that's it.
This is more flexible since it allows us to create different PipelineState for different render-passes. This is very efficient for D3D12, not so much for D3D11 and OpenGL. I will handle it when it becomes a pain (in a backward compatible way).

Vao (which I'd like to rename) now holds the primitive topology. It basically defines the geometric properties of the object. Not sure about that, but I didn't see a case yet where we used different topologies with the same buffers.