SceneEditor is working.
- Had to work around a defaultVS issue - the input layout must match the VS input signature. I'm using defines for now, but I'll probably want to procedurally generate the default VS (or parts of it).
- Fixed the SceneEditor GUI window size and initial position
- Group all the light sources under a "Lights" tab

The SceneEditor kind-of works. Removing models will crash the application. The problem is that the driver doesn't do any resource tracking anymore. The RenderContext defers execution after all the commands have been recorded - including the GUI - and this causes some resources to be released prematurely. We need to defer the destruction of resources to the end of the frame. Yay DX12!