Removed the SceneRenderer map from the ModelRenderer.
For now I will just create a new Scene and SceneRenderer everytime a draw is called. I try a few scenes, I didn't notice any performance issues. In any cases, ModelRenderer should not be used if one wants performance.

The problem was that the map was static. When the model was released by the user, it was still part of the map(via the Scene and the SceneRenderer) - resulting in memory leak and crash when the application exists. There are ways to solve it, but it's not worth the effort.