Fixed crash at exit.
It was caused by the static VB in the FullScreenPass. Because it was static, it was destroyed very late after most things were gone (including the device).
I fixed that by adding a static reference count for the number of allocated objects and release the VB when the reference count reaches zero.

No more crashes (at least, none that I know of :) )