Added clear function for texture UAVs
It's not fully featured - you can only clear the entire resource.
There's always the option to specify the view params to the clear function, but I don't like it.
Option 2 is to let the clear function get the view. We will need to store the resource pointer inside the view and probably create a clearUav when we create the regular UAV. I don't like that to.
I'll think about it.