Hacking to hook up camera parameters.
I've decided that for now the existing `Falcor::ConstantBuffer` type is a good stand-in for a component instance (even if it isn't backed by a descriptor table right now).
The way Falcor is designed, a "constant buffer" can actually include not just uniform data, but also textures, samplers, etc.
This lets me re-use a lot of existing infrastructure instead of creating things from scratch (and it hopefully means we can swap in descriptor-table use without changing the application code too much).

Right now the `Camera` class relies on other code to pass in an appropriate Spire context to use for looking up the component class, eventually that will hopefully just use the global/singleton context that everything else uses.

Also, there is some duplication where the reflection data for a component is getting computed redundantly; eventually it would be good to have Falcor maintain a cache that it looks up using a `SpireModule*` that then holds its reflection data for the component class.

I'm not very happy with how the Falcor `BufferReflection` is factored, since it basically munges together the type-specific and variable-specific bits of data (e.g., the binding index should belong to the parent program, while the contained variables would belong to the component that gets plugged in there). That can be addressed later.