Make `Material` a shader interface, so we generate some variants.
This change modifies the "shader library" so that there is a base `Material` interface with concrete implementations `ConstantMaterial` and `TexturedMaterial`.
We then select between these using CPU-side logic in the `Falcor::Material` class.
The important result here is that we are able to execute the correct specialized shader variant (and pass things through the correct specialized root signature) without the application code needing to be aware.