Refactor `ComponentInstance` code.
A lot of what `ComponentInstance` does is stuff `ProgramVars` used to do (and now `ProgramVars` delegates a bunch of calls to its first component).
This change moves a lot of component-instance code into ProgramVars.cpp, so that it will hopefully produce decently readable diffs when compared against what used to be there.

This also deletes a bunch of code from `ProgramVars` that was made dead by the switch to Spire.

As a side effect of this cleanup, it should now be possible for more existing Falcor application code to Just Work with Spire (once their shaders are ported).
The main sticking point is (and will continue to be) any code related to setting uniforms in the "main" constant buffer of a given shader.