Various fixes for Spire usage
- The main fix was to remove an `if` from `ComponentInstance::getApiHandle` that was trying to handle `NULL` shader resource views. Unfortunately, that logic was skipping the increment of the descriptor index, and so we weren't laying things out correctly. The correct logic is to substitute in a dummy SRV, and Falcor already has code to supply one for us.

- The other big fix is to re-use the reflection data for Spire components, which in turn means that we don't end up creating new program variants all the time (we were using the reflection data pointer as the key for computing the variant).
  - A simpler fix might have been to use the `SpireModule*` as the key instead of the generated data, but it was also worthwhile to fix things so that we don't re-generate reflection data all the time
  - There are still redundant PSOs getting created, but this seems to be due to some other broken-ness in how Falcor is doing things (and not directly Spire's fault). We may want to fix the underlying issue, but then we have to ask whether we get to count it in our favor for evaluation.

At this point the "ogre" model at least loads and renders with textures, but it appears that its other material-related uniform parameters are all out of whack. I will investigate that next.

Note: one unfortunate thing I've found is that somewhere along the way, it has become impossible to reliably use the built-in Visual Studio graphics debugging tools with Falcor+Spire. It seems like it is incorrectly reporting the identity of shaders that get used, and thus not providing the right assembly when I try to step through them.