Fix caching of program versions for Spire.
The `Program` class was caching `ProgramVersion`s in a dictionary keyed on the "define list," when for Spire we really want it to be keyed on the list of active components.

Actually... we want it to be keyed on both (for now - at least until we can remove all uses of `#if` stuff in shader code), so I've gone ahead and made the key to the "variant cache" be composed of both the define list and component-class list.