Put constant buffers for component instances in root table slots.
I'd previously said that when a component has no resource parameters, we might as well put its constant buffer directly in a root-table slot rather than nested in a descriptor table.
This change just makes that same optimization unconditionally.

The main benefit of doing this is that it obviates the need for some fiddly logic we had for dealing with invalidation of the resource desriptor table whenever a constant-buffer write gets performed.

There doesn't appear to be a major performance benefit of doing this, however, for the current workload I'm testing.