Fix binding indices for resources/samplers in cbuffers.
The binding index info for a variable indside a buffer is always its offset relative to the *start* of the buffer (which may have its own texture binding slot, etc.).
As a result, we need to be sure to add the base index from the buffer into the offset for any variable inside it.

This fix requires some companion changes to the Spire reflection API, and also points out that the public API really needs to unify the notion of a refleciton "parameter" with a reflection "variable," since the distinction is pretty much unimportant to users.