Amortize Spire compiler load time.
This involved a major revamp to the public Spire API (long overdue) to make it easy to amortize out the main startup cost (loading the Spire stdlib).
I'd still like to speed that step up, but ast least at this point the rewriter step takes less time than the downstream fxc compilation (as it should).