Use Spire to enable hot reload.
The main change here is using a Spire-provided API to query the list of file paths that a given `ProgramVersion` compilation depended on.
Note that this means the equivalent info is no longer available at the `Shader` level.

This also includes some unrelated changes to track Spire API changes I had to make to keep the main Spire compiler working.