Fixed an issue with the state graph
We can't reset the graph at the start of the frame. The root data is not equivalent to the current state of the graph. If we do that, the tracking basically meaningless (even though in practice it actually works most of the time).
The correct solution is to just use the graph as a *graph*, not as a tree. When a new GSO is required, we first scan the entire graph to see if there is already an existing node which matches the current state. If there is, we add an edge to that node. Otherwise we create a new state object.
Unlike the previous implementation, this allows to ignore set*() calls which result in NOP.