Fix support for querying the size of a uniform array.
Falcor application code used to query the size of a uniform array, by using a call that looks like it is querying the size of the first element:

    getVariableData("someArrayVar[0]")->arraySize

In the Spire-based reflection data, that returns the array size of the type of `someArrayVar[0]` and not of `someArrayVar` itself.
This behavior actually seems like what we want long-term, and the existing Falcor code has just been adding `"[0]"` to the end of various queries to get around the way that the Microsoft compiler doesn't always provide reflection for the array itself (just the individual elements).

The right query would of course be:

    getVariableData("someArrayVar")->arraySize

but that wasn't implemented in the current code that maps Spire reflection over to Falcor.
This change adds the missing support.

I also removed support for the `allowNonIndexedArray` parameter to `getVariableData()` because it doesn't seem like something we want or need long-term.
I left the old declarations in place (with a deprecation attribute) and routed calls to the new versions.

In order to support deprecation attributes, I added a macro that is supposed to abstract the compiler-specific behavior (although I haven't checked on the gcc/clang syntax for deprecation).