Fix a bug in how rewriter is invoked.
The existing code was assuming that the index of a shader string in `Program::mShaderStrings` will be equal to the index of the "translation unit" as understood by the Spire API.
This is not true in general, because there can be empty entries in `mShaderStrings` that get skipped, and it only worked because all the shaders I'd tested so far only used VS/PS which are the first two entries.

This change fixes the issue by tracking separate counters for the translation-unit index as needed.