Move Spire logic out of `createProgramVersion`
The `Program::createProgramVersion` function is declared `virtual` so that specialized subclasses of `Program` can override it (e.g., to deal with alternative pipelines).
By putting all the Spire-related logic in the default implementation, we make it harder for such subclasses to take advantage of the rewriter (which is now our only way of doing preprocessing).

This change moves the rewriter logic (and also the tracking of dependencies for hot reload) out to a helper function `preprocessAndCreateProgramVersion`, which then calls through to `createProgramVersion` if preprocessing went well.
There is only one existing call site to `createProgramVersion` so updating it was relatively easy.

Because `createProgramVersion` now needs the output of preprocessing (including the reflection data it extracts), I went ahead and stored those outputs in the `Program` for now, as `mutable` values.
This is a worse design in most respects, but it doesn't require as many changes to clients of the current API.
To help keep things clear, I changed the storage for shader strings to differentiate the "original" strings from the "preprocessed ones."

A later cleanup pass should change this design so that the results of preprocessing are passed into `createProgramVersion` explicitly.