Started working on the intrefaces for the new classes.
I need to figure out if DescriptorPool is a good replacement for DescriptorHeap. It doesn't look like that - it actually look like a pool is a collection of heaps (depending on the underlying API).
If so, is it redundant? For now it looks like we need both - that will simplify the API abstraction - but it feels like over-design.

To be continued