Add support for loading custom builtins for Spire rewriter.
The `addSpireBuiltins()` function can be used to load additional code as if it had been declared in the Spire "standard library" so that all shaders that get loaded after that can use those types/functions. For right now this is a bit limited, because certain core things are defined internal to the compiler rather than via stdlib, but I plan to change that over time.