Fix ShaderToy example to not (accidentally) use global-scope uniform
The declaration:

    vec3 iMouse = vec3(0, 0, 0);

is interpreted by GLSL as declaring a "thread-local" variable. If you want this to be a uniform parmameter you need to mark it explicitly with `uniform`.

HLSL ends up working the other way around. The declaration of `iMouse` above creates a uniform at global scope (which will end up in a magic constant buffer called `$Globals`), and you need to use an explicit `static` keyword if you want the "thread-local" behavior.

Right now Falcor doesn't support global-scope uniforms, and the Spire rewriter's support is only partial, so the ShaderToy app was failing because of this declaration.

But of course ShaderToy actually *wanted* the "thread-local" case and not the uniform, so the simple fix was to change it to:

    #ifdef FALCOR_HLSL
    static
    #endif
    vec3 iMouse = vec3(0, 0, 0);