Vulkan: Add Vulkan SDK files to `Externals`
This should just cover the files strictly needed to build for Vulkan, and does *not* include the entire SDK (some 600+ MB).
It also ignores any headers that aren't actually included by a baseline Vulkan app, and definitely exclused the C++ wrapper/helper interfaces (since that is a 1MB `.hpp` file!).