Slang: Add slang source to repository.
Similar to what we did when this was called "Spire," we are not using submodules because they create headaches for the user.
Instead, we are just adding the relevant source code to the repository directly.

Unlike with the "Spire" code, we are carefully just adding the stuff we directly need, and not *everything*.