Convert `ShaderUtils` part of the shader library to Slang
The easy part here was to rename, e.g., `Shading.h` over to `Shading.slang` and then change all `#include "Shading.h"` over to `__import Shading;`.
Some dependencies between those files had to be made more explicit, now that they cannot depend on order of includes to make code visible.

The hard part here was what to do with `HostDeviceData.h`, because that file contains both a bunch of macros that downstream code needs (and which *wouldn't* be imported via `__import`) along with definitions that are used downstream (and which would end up with duplicate definitions if we `#include` those definitions into multiple distinct modules).

The solution implemented here is:

- The main code of `HostDeviceData.h` is split between `HostDeviceSharedMacros.h` which defines the macros, and `HostDeviceSharedCode.h` which defines the actual code

- When included from host code, `HostDeviceData.h` will just `#include` both `HostDeviceSharedMacros.h` and `HostDeviceSharedCode.h`, leading to the same result as today.

- When included from shader code, `HostDeviceData.h` will `#include "HostDeviceSharedMacros.h"` so the macros are in scope, but `__import HostDeviceSharedCode;` to ensure that the actual definitions are imported properly.

This means that both user code and the Slang-based library code can just `#include "HostDeviceData.h"` and get the behavior they expect.