RootSignature works only with DescriptorSet.
The logic for generating the signature from a reflection object is API specific. Vulkan doesn't support root-descriptors which DX does. The reason for that is that using CBV as a root-descriptor results in better performance then using a descriptor-table.
This is probably a CPU overhead issue which I will need to tackle at some point, but for now using a root-desc just works better.