Remove HLSL/GLSL compatibility header
The key change here is the removal of `HlslGlslCommon.h`, which had been used to make a certain number of GLSL types available in HLSL code.
This is largely a byproduct of Falcor's legacy, starting as a GL-only framework.
Now that all of the shader code has been ported to HLSL, and cross-API compatibility will be handled via translation of that HLSL input, it makes sense to start writing things using more idiomatic pure HLSL.

The remaining changes here are just renaming changes in all of the affected shader code: `vec3` becomes `float3`, etc.

The one notable exception to all this is `toy.fs` for the ShaderToy example, since the whole point there is that it is supposed to be GLSL code.
I've gone ahead and hand-copied a few definitions into `toyContainer.fs` just to make things work, but this isn't really a robust solution for arbitrary GLSL code.