Switch HLSL matrix convention.
Up to this point, the Falcor code has been using a convention for matrices as folows:

- HLSL `floatNxM` is assumed equivalent to GLSL `matNxM`
- HLSL `m[i]` == GLSL `m[i]`
- HLSL `mul(m, v)` == GLSL `m * v`
- D3D and GLSL both use default matrix memory layout conventions in constant buffers

Unfortunately, these choices can't all be kept consistent if `N` is not equal to `M` (that is, for non-square matrices).
This is because from a typing standpoint, `mul(m,v)` is equivalent to `v * m` (note the reversed order).

(This all stems from the fact that `floatNxM` is officialy `N` rows and `M` columns, while `matNxM` is officialy `N` columns and `M` rows. This mostly doesn't come up as an issue, until you get to matrix-vector and matrix-matrix multiply, where the definitions of "rows" and "columns" suddenly become important.)

All of this is coming to a head because we need to support cross-compilation of Slang code to GLSL, and Slang follows the HLSL conventions.

Given this situation, there are only a few viable choices for how to get out of the mess:

1. Just give up on non-square matrices and hand-wave the whole thing. This might be viable, but it is unatractive as a long-term solution.

2. Keep `mul(m,v)` == `m * v`, and thus decide that `floatNxM` is equivalent to `matMxN` (transposed dimensions).
  - This means that HLSL/Slang `m[i]` turns into some messy expression in GLSL (that isn't an l-value).
  - I'm not sure what this would mean for memory layout conventions (this stuff is confusing)

3. Decide that `mul(m,v)` == `v * m`, which means we can keep `floatNxM`==`matNxM` and `m[i]`==`m[i]`

Slang already implements option (3) as a matter of policy (it is pretty much the only viable option for a cross-compiler).
For what it is worth, I believe that the glslang compiler implements the same policy when it comes to accepting HLSL input (although to be fair, I helped influence that policy decision...).

Implementing (3) in Falcor has two consequences, which this change tries to address:

- First, we need to set the matrix layout conventions to the non-default value in order to ensure that matrix data loaded from the host (which always has the same layout, independent of API) will be interpreted "transposed" by HLSL.

- Second, we need to change most `mul(a,b)` calls in HLSL or Slang code to `mul(b,a)` to deal with the change
  - I say "most" because there will be times where a matrix is constructed on the fly (not passed in from the host), and in those cases the existing HLSL code will continue to have its old meaning - changing the packing rules for constant buffers doesn't change how local variables are stored.