Fix gaussian blur kernel to use Slang compile-time loop
Both HLSL and GLSL require that the texel offset in a fetch-with-offset call be a compile-time constant, but HLSL is willing to go to heroics to ensure this, while GLSL will often just given an error.
In particular, putting an `[unroll]` on a loop gives no strong guarantees, because GLSL doesn't have that construct.

New Slang changes will add a compile-time loop construct, and these are shader changes designed to take advantage of it.
Using the looping construct is only part of the fix, since we also need to avoid passing things as function parameters that we need to be visible compile-time-constant.