Wait on an fence when requesting the next image to make sure it's ready
I didn't notice any performance impact vs using a semaphore.
The problem with using a semaphore is that we can potentially get to a point when we are submitting many cmd-buffers and the presentation engine can't keep up. Theoretically, it will lead to an increased latency between the UI and the presented image.