Fix Slang shader code for AMD
The root cause here is the interaction of two bad choices:

1. The Falcor shader library has a bit of a fetish for marking things (including shader parameters) as `const` whenever possible

2. The Slang compiler currently isn't smart enough to know the difference between GLSL `const` and HLSL `const`

This meant that we passed through GLSL code that declared a bunch of things `const` that didn't need to be declared that way. GLSL has historically had a very stringent idea of what `const` means, and while it seems that the current language spec is more lax, it may be the case that some drivers get confused by our code applying `const` to things that are not compile-time constants.

For now, I'm proposing to fix this by banning `const` on parameters from the Falcor shader library. A better long-term fix is for Slang to not pass through HLSL/Slang `const` directly to GLSL.